Extract chat attestation helpers
