+++ js/chat.js
@@ -33,6 +33,7 @@ import {
   attachLensSources, buildChatSystemPrompt, buildMultiPersonaInstruction,
   buildPersonalityPrompt, buildTaggedChatMessages, buildWebSearchHint,
 } from './chat-prompt-context.js';
+import { e2eeLockFootnote, e2eeLockHTML } from './chat-attestation.js';
 
 // ═══════════════════════════════════════════════
 // ABORT CONTROLLER (stop streaming)
@@ -357,37 +358,6 @@ export function updateSummaryButton() {
   const hasSummary = !!thread?.summary;
   btn.classList.toggle('has-summary', hasSummary);
   btn.title = hasSummary ? 'View summary' : 'Summarize this conversation';
-}
-
-/** Build attestation tooltip text. */
-function _attestationTooltip(attestation) {
-  const ok = attestation.nonceVerified && attestation.signingKeyBound && !attestation.debugMode;
-  const lines = [
-    `Nonce: ${attestation.nonceVerified ? '\u2713' : '\u2717'}`,
-    `Key binding: ${attestation.signingKeyBound ? '\u2713' : '\u2717'}`,
-    `Debug mode: ${attestation.debugMode ? 'YES \u2717' : 'no \u2713'}`,
-    attestation.serverTdxValid != null ? `Server TDX: ${attestation.serverTdxValid ? '\u2713' : '\u2717'}` : null,
-    attestation.dcap ? `DCAP: ${attestation.dcap.status}` : null,
-  ].filter(Boolean);
-  return (ok ? 'TEE attestation verified' : 'TEE attestation FAILED') + '\n' + lines.join('\n');
-}
-
-/** E2EE lock HTML for header: 🔒 alone, or 🔒 + colored ✓/✗ with tooltip. */
-function e2eeLockHTML(attestation) {
-  if (!attestation) return ' \uD83D\uDD12';
-  const ok = attestation.nonceVerified && attestation.signingKeyBound && !attestation.debugMode;
-  const color = ok ? '#22c55e' : '#ef4444';
-  const mark = ok ? '\u2713' : '\u2717';
-  return ` <span title="${_attestationTooltip(attestation)}">\uD83D\uDD12<span style="color:${color};font-weight:bold">${mark}</span></span>`;
-}
-
-/** E2EE lock HTML for cost footnotes. */
-function e2eeLockFootnote(attestation) {
-  if (!attestation) return ' \u00b7 \uD83D\uDD12 e2ee';
-  const ok = attestation.nonceVerified && attestation.signingKeyBound && !attestation.debugMode;
-  const color = ok ? '#22c55e' : '#ef4444';
-  const mark = ok ? '\u2713' : '\u2717';
-  return ` \u00b7 <span title="${_attestationTooltip(attestation)}">\uD83D\uDD12<span style="color:${color};font-weight:bold">${mark}</span> e2ee</span>`;
 }
 
 // Auto-refresh header when attestation becomes available
